Move getTypeContent from util/content to util/prisma
