--- backend/src/util/content.ts
@@ -1,5 +1,0 @@
-import { Prisma } from "@prisma/client";
-
-export function getTypedContent<T>(content: T & { config: Prisma.JsonValue }) {
-  return { ...content, config: content.config as any };
-}
